Make assertion failures more descriptive in `BalanceSpec`.
The `BalanceSpec` module contains several properties that make multiple
internal assertions. When one of them fails, it's rather difficult to
determine exactly which one has failed.

This commit adds an `assertWith` function, making it possible to provide
a description with each assertion that is printed out when the assertion
fails.